data.europa.eu/requirements.txt
psycopg2-binary
python-dotenv
rdflib
requests
python-jsonpath

